refactor: collapse Fintual ingestion pipeline into FintualPerformance
